fix: remove `RustTarget` use single ShaSum impl
