+++ crates/fluvio-hub-util/src/fvm/api/download.rs
@@ -1,33 +1,19 @@
 //! Download API for downloading the artifacts from the server
 
-use std::path::{PathBuf};
-use std::io::{Cursor, copy, Read};
+use std::path::PathBuf;
+use std::io::{Cursor, copy};
 use std::fs::File;
 
 use anyhow::{Error, Result};
 use http_client::async_trait;
-use sha2::{Sha256, Digest};
 use surf::StatusCode;
 
 use crate::fvm::Artifact;
-
-/// Generates the Sha256 checksum for the specified file
-fn shasum256(file: &File) -> Result<String> {
-    let meta = file.metadata()?;
-    let mut hasher = Sha256::new();
-    let mut buffer = vec![0; meta.len() as usize];
-
-    file.read_exact(&mut buffer)?;
-    hasher.update(buffer);
-
-    let output = hasher.finalize();
-    Ok(hex::encode(output))
-}
+use crate::utils::sha256_digest;
 
 /// Verifies downloaded artifact checksums against the upstream checksums
 async fn checksum(artf: &Artifact, path: &PathBuf) -> Result<()> {
-    let file = File::open(path)?;
-    let local_file_shasum = shasum256(&file)?;
+    let local_file_shasum = sha256_digest(path)?;
     let upstream_shasum = surf::get(&artf.sha256_url)
         .await
         .map_err(|err| Error::msg(err.to_string()))?
@@ -121,8 +107,7 @@ mod test {
         artifact.download(target_dir.clone()).await.unwrap();
 
         let binary_path = target_dir.join("fluvio");
-        let file = File::open(binary_path).unwrap();
-        let downstream_shasum = shasum256(&file).unwrap();
+        let downstream_shasum = sha256_digest(&binary_path).unwrap();
         let upstream_shasum = surf::get(&artifact.sha256_url)
             .await
             .map_err(|err| Error::msg(err.to_string()))
@@ -133,54 +118,5 @@ mod test {
             .unwrap();
 
         assert_eq!(downstream_shasum, upstream_shasum);
-    }
-
-    #[test]
-    fn creates_shasum_digest() {
-        use std::fs::write;
-
-        let tempdir = TempDir::new().unwrap().into_path().to_path_buf();
-        let foo_path = tempdir.join("foo");
-
-        write(&foo_path, "foo").unwrap();
-
-        let foo_a_checksum = shasum256(&File::open(&foo_path).unwrap()).unwrap();
-
-        assert_eq!(
-            foo_a_checksum,
-            "2c26b46b68ffc68ff99b453c1d30413413422d706483bfa0f98a5e886266e7ae"
-        );
-    }
-
-    #[test]
-    fn checks_files_checksums_diff() {
-        use std::fs::write;
-
-        let tempdir = TempDir::new().unwrap().into_path().to_path_buf();
-        let foo_path = tempdir.join("foo");
-        let bar_path = tempdir.join("bar");
-
-        write(&foo_path, "foo").unwrap();
-        write(&bar_path, "bar").unwrap();
-
-        let foo_checksum = shasum256(&File::open(&foo_path).unwrap()).unwrap();
-        let bar_checksum = shasum256(&File::open(&bar_path).unwrap()).unwrap();
-
-        assert_ne!(foo_checksum, bar_checksum);
-    }
-
-    #[test]
-    fn checks_files_checksums_same() {
-        use std::fs::write;
-
-        let tempdir = TempDir::new().unwrap().into_path().to_path_buf();
-        let foo_path = tempdir.join("foo");
-
-        write(&foo_path, "foo").unwrap();
-
-        let foo_a_checksum = shasum256(&File::open(&foo_path).unwrap()).unwrap();
-        let foo_b_checksum = shasum256(&File::open(&foo_path).unwrap()).unwrap();
-
-        assert_eq!(foo_a_checksum, foo_b_checksum);
     }
 }
